rootfs/requirements.txt
# Drycc manager requirements
authlib==0.15.5
django==4.1
pytz==2022.2.1
django-cors-headers==3.10.0
django-guardian==2.4.0
djangorestframework==3.12.4
gunicorn==20.1.0
uvicorn==0.18.2
asgiref==3.5.2
psycopg2-binary==2.9.2
requests==2.26.0
requests-toolbelt==0.9.1
django_redis==5.0.0
hiredis==2.0.0
dj-database-url==0.5.0
social-auth-app-django==5.0.0
python-jose==3.3.0
whitenoise==5.3.0
celery==5.2.1
stripe==3.0.0